requirements.txt
boto3
botocore
flask
flask_cors
numpy
opencv-python
Pillow
scikit-learn
